refactor: rename 'BaseFunction.mts' to reflect non-serializable functions
